Map location set to device in baseline model load
I use my QA projects across devices and a problem encountered was deserialization if the GPU/device used for creating the baseline model was different to the device used for further training from baseline specifically, where I had trained baseline on a device with multiple GPUs and was retraining on a laptop with a single GPU. In training of baseline, I had selected GPU 1 in the QuickAnnotator config and in retraining, I was using GPU 0. 
OS (Ubuntu 20.04), NVidia drivers, conda environment etc. were identical between systems. 
The error message encountered was identical to the one outlined here: https://stackoverflow.com/questions/56369030/runtimeerror-attempting-to-deserialize-object-on-a-cuda-device  where users of PyTorch attempted to retrain models using CPU instead of GPU. The error was traced to line 169 of this file (train_model.py): the torch.load() line.
As outlined in the above stackoverflow, the solution is through unambiguous assignment of the device to load the model to through the map_location argument with the PyTorch load() function. By retrieving the device variable one line earlier and unambiguously mapping location in line 169, this error was resolved with no other changes in behaviour.